pyproject.toml
[tool.poetry]
name = "steam_items_api"
version = "1"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.10.5"
requests = "^2.32.3"
starlette = "^0.38.5"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
